defrag: shrink WASM binary 26% and maximize shared code between native/WASM builds
- Add [profile.release] with lto, codegen-units=1, strip, opt-level=z
- Gate wast crate as native-only (remove from WASM build)
- Gate docs.rs search/list functions as native-only
- Create shared diagnostics_core/tree_sitter.rs for both targets
- Create shared features/signature/call_info.rs for both targets
- Promote features/signature from native-only to shared module
- Remove --no-opt from wat-lsp package build to enable wasm-opt
- Move playground compile button into editor tabs
