+++ Cargo.toml
@@ -42,6 +42,7 @@ native = [
     "dep:tree-sitter-language",
     "dep:dashmap",
     "dep:clap",
+    "dep:wast",
 ]
 wasm = [
     "dep:wasm-bindgen",
@@ -58,7 +59,7 @@ wasm = [
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 regex = "1.10"
-wast = "243.0"
+wast = { version = "243.0", optional = true }
 
 # Native-only dependencies
 tower-lsp = { version = "0.20", optional = true }
@@ -92,3 +93,16 @@ required-features = ["native"]
 name = "parser"
 harness = false
 required-features = ["native"]
+
+# ============================================================================
+# Release profile — tuned for small WASM binary + fast native binary
+# ============================================================================
+
+[profile.release]
+lto = true
+codegen-units = 1
+strip = true
+
+# WASM-specific: optimise for binary size
+[profile.release.package.wat-lsp-rust]
+opt-level = "z"
