+++ src/diagnostics/tree_sitter_diagnostics.rs
@@ -1,79 +1,22 @@
-use crate::utils::node_to_range;
-use tower_lsp::lsp_types::*;
-use tree_sitter::{Node, Tree};
+use tower_lsp::lsp_types::Diagnostic;
+use tree_sitter::Tree;
 
-/// Provide diagnostics for syntax errors in the document
+/// Provide diagnostics for syntax errors in the document.
+///
+/// Delegates to the shared `diagnostics_core::tree_sitter` implementation and
+/// converts core diagnostics to tower-lsp types.
 pub fn provide_tree_sitter_diagnostics(tree: &Tree, source: &str) -> Vec<Diagnostic> {
-    let mut diagnostics = Vec::new();
-    walk_tree_for_errors(tree.root_node(), source, &mut diagnostics);
-    diagnostics
-}
-
-/// Recursively walk the tree and collect ERROR nodes as diagnostics
-fn walk_tree_for_errors(node: Node, source: &str, diagnostics: &mut Vec<Diagnostic>) {
-    if node.kind() == "ERROR" {
-        let diagnostic = create_error_diagnostic(node, source);
-        diagnostics.push(diagnostic);
-    }
-
-    // Check for MISSING nodes as well (incomplete syntax)
-    if node.is_missing() {
-        let diagnostic = create_missing_diagnostic(node, source);
-        diagnostics.push(diagnostic);
-    }
-
-    // Recursively check children
-    let mut cursor = node.walk();
-    for child in node.children(&mut cursor) {
-        walk_tree_for_errors(child, source, diagnostics);
-    }
-}
-
-/// Create a diagnostic for an ERROR node
-fn create_error_diagnostic(node: Node, source: &str) -> Diagnostic {
-    let range: Range = node_to_range(&node).into();
-    let text = &source[node.byte_range()];
-
-    let message = if text.trim().is_empty() {
-        "Syntax error: unexpected token".to_string()
-    } else {
-        format!("Syntax error near: {}", text.lines().next().unwrap_or(text))
-    };
-
-    Diagnostic {
-        range,
-        severity: Some(DiagnosticSeverity::ERROR),
-        code: None,
-        code_description: None,
-        source: Some("wat-lsp".to_string()),
-        message,
-        related_information: None,
-        tags: None,
-        data: None,
-    }
-}
-
-/// Create a diagnostic for a MISSING node
-fn create_missing_diagnostic(node: Node, _source: &str) -> Diagnostic {
-    let range: Range = node_to_range(&node).into();
-
-    Diagnostic {
-        range,
-        severity: Some(DiagnosticSeverity::ERROR),
-        code: None,
-        code_description: None,
-        source: Some("wat-lsp".to_string()),
-        message: format!("Missing {}", node.kind()),
-        related_information: None,
-        tags: None,
-        data: None,
-    }
+    crate::diagnostics_core::tree_sitter::provide_tree_sitter_diagnostics(tree, source)
+        .into_iter()
+        .map(Diagnostic::from)
+        .collect()
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::tree_sitter_bindings::create_parser;
+    use tower_lsp::lsp_types::DiagnosticSeverity;
 
     #[test]
     fn test_no_errors_in_valid_code() {
@@ -101,7 +44,6 @@ mod tests {
             "Invalid code should have diagnostics"
         );
 
-        // Check that at least one diagnostic is an error
         assert!(
             diagnostics
                 .iter()
@@ -117,8 +59,6 @@ mod tests {
         let tree = parser.parse(document, None).unwrap();
 
         let diagnostics = provide_tree_sitter_diagnostics(&tree, document);
-        // Incomplete code may or may not produce errors depending on grammar
-        // Just ensure we don't crash
         let _ = diagnostics;
     }
 
@@ -131,7 +71,6 @@ mod tests {
         let diagnostics = provide_tree_sitter_diagnostics(&tree, document);
         assert!(!diagnostics.is_empty());
 
-        // Verify that diagnostics have valid ranges
         for diagnostic in &diagnostics {
             assert!(diagnostic.range.start.line <= diagnostic.range.end.line);
             if diagnostic.range.start.line == diagnostic.range.end.line {
@@ -231,7 +170,6 @@ mod tests {
 
     #[test]
     fn test_store8_load8_valid() {
-        // These are valid WASM instructions and should not produce errors
         let document = r#"(module
   (memory 1)
   (func $test
