+++ src/features/signature/tests.rs
@@ -1,3 +1,4 @@
+use super::call_info::{extract_name_from_call, find_function_call, CallInfo};
 use super::*;
 use crate::features::test_utils::{create_signature_test_symbols, create_test_tree};
